+++ src/common/rabbitmq/rabbitmq.token.handler.service.ts
@@ -1,0 +1,51 @@
+import { Inject, Injectable } from '@nestjs/common';
+import { CacheInfo } from '../../utils/cache.info';
+import { NotifierEvent } from './entities/notifier.event';
+import { BinaryUtils, CachingService } from '@elrondnetwork/erdnest';
+import { OriginLogger } from '@elrondnetwork/erdnest';
+import { EsdtService } from 'src/endpoints/esdt/esdt.service';
+import { ClientProxy } from '@nestjs/microservices';
+
+@Injectable()
+export class RabbitMqTokenHandlerService {
+  private readonly logger = new OriginLogger(RabbitMqTokenHandlerService.name);
+
+  constructor(
+    private readonly cachingService: CachingService,
+    private readonly esdtService: EsdtService,
+    @Inject('PUBSUB_SERVICE') private clientProxy: ClientProxy,
+  ) { }
+
+  public async handleTransferOwnershipEvent(event: NotifierEvent): Promise<boolean> {
+    const tokenIdentifier = BinaryUtils.base64Decode(event.topics[0]);
+
+    try {
+      const esdtProperties = await this.esdtService.getEsdtTokenPropertiesRaw(tokenIdentifier);
+      if (!esdtProperties) {
+        return false;
+      }
+
+      await this.invalidateKey(
+        CacheInfo.EsdtProperties(tokenIdentifier).key,
+        esdtProperties,
+        CacheInfo.EsdtProperties(tokenIdentifier).ttl
+      );
+
+      return true;
+    } catch (error) {
+      this.logger.error(`An unhandled error occurred when processing transferOwnership event for token with identifier '${tokenIdentifier}'`);
+      this.logger.error(error);
+
+      return false;
+    }
+  }
+
+  private async invalidateKey(key: string, data: any, ttl: number) {
+    await this.cachingService.setCache(key, data, ttl);
+    this.refreshCacheKey(key, ttl);
+  }
+
+  private refreshCacheKey(key: string, ttl: number) {
+    this.clientProxy.emit('refreshCacheKey', { key, ttl });
+  }
+}
